# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Backend Coding Test API Document",
    "description": "Document for API of this project",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8010",
  "basePath": "/",
  "tags": [
    {
      "name": "API_Health",
      "description": "API for health check"
    },
    {
      "name": "Rides",
      "description": "API for rides"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "tags": ["API_Health"],
        "summary": "API Health Check",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Health"
            }
          }
        }
      }
    },

    "/rides": {
      "get": {
        "tags": ["Rides"],
        "summary": "API Get Rides",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "$ref": "#/definitions/Rides"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/GeneralError"
            }
          }
        }
      },
      "post": {
        "tags": ["Rides"],
        "summary": "API Create Rides",
        "parameters": [
          {
            "name": "ride",
            "in": "body",
            "description": "Ride information that we want to create",
            "schema": {
              "$ref": "#/definitions/RideParametter"
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Ride"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/GeneralError"
            }
          }
        }
      }
    },
    "/rides/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of ride that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Rides"],
        "summary": "API Get Deatil of Ride",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Ride"
            }
          },
          "400": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/GeneralError"
            }
          }
        }
      }
    }
  },
  "schemes": ["http"],
  "consumes": ["application/json", "text"],
  "produces": ["application/json", "text"],
  "definitions": {
    "Health": {
      "type": "string"
    },
    "Ride": {
      "properties": {
        "rideID": {
          "type": "string"
        },
        "startLat": {
          "type": "number",
          "format": "double"
        },
        "startLong": {
          "type": "number",
          "format": "double"
        },
        "endLat": {
          "type": "number",
          "format": "double"
        },
        "endLong": {
          "type": "number",
          "format": "double"
        },
        "riderName": {
          "type": "string"
        },
        "driverName": {
          "type": "string"
        },
        "driverVehicle": {
          "type": "string"
        },
        "created": {
          "type": "string"
        }
      }
    },
    "Rides": {
      "type": "array",
      "$ref": "#/definitions/Ride"
    },
    "GeneralError": {
      "properties": {
        "error_code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "RideParametter": {
      "properties": {
        "start_lat": {
          "type": "number",
          "format": "double"
        },
        "start_long": {
          "type": "number",
          "format": "double"
        },
        "end_lat": {
          "type": "number",
          "format": "double"
        },
        "end_long": {
          "type": "number",
          "format": "double"
        },
        "rider_name": {
          "type:": "string"
        },
        "driver_name": {
          "type:": "string"
        },
        "driver_vehicle": {
          "type": "string"
        }
      }
    }
  }
}